Refactoring and config load/check
