+++ client/.gitignore
@@ -2,3 +2,4 @@ client
 client-*
 *.pem
 *.key
+config.json
